+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Hide accessibility settings
 - Fix quickicons markup (thanks @Quy)
 - Fix top nav dropdown header styling
+- Fix missing WebAuthn support for login
 
 ## 0.0.15
 - Add search filter to "Add Module" modal